second rectification

+++ python_introduction/finance_calculator.py
@@ -1,13 +1,13 @@
 #Prompt the user for their monthly income and expenses
-income = float(input("Enter your monthly income: "))
-expenses = float(input("Enter your total monthly expenses: "))
+income = int(input("Enter your monthly income: "))
+expenses = int(input("Enter your total monthly expenses: "))
 
 #Calculate monthly savings
 monthly_savings = income - expenses
 
 #Annual savings projection
 
-projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 
 #output results
 print(f"Projected savings after one year, with interest, is: {projected_savings}")
